Renamed core to classes
