+++ src/transformers/generation/continuous_batching/__init__.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 from .cache import PagedAttentionCache
 from .continuous_api import ContinuousBatchingManager, ContinuousMixin
-from .core import RequestState, RequestStatus
+from .classes import RequestState, RequestStatus
 
 
 __all__ = ["PagedAttentionCache", "RequestState", "RequestStatus", "ContinuousMixin", "ContinuousBatchingManager"]
