+++ src/transformers/generation/continuous_batching/classes.py
